+++ playroom/snippets.tsx
@@ -464,10 +464,6 @@ const headerSnippets: Array<Snippet> = [
             header={
                 <Header
                     title="The last invoice is available"
-                    preamount="Some text (text)"
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle="Subtitle"
                 />
             }
             extra={<Placeholder />}
@@ -483,10 +479,6 @@ const headerSnippets: Array<Snippet> = [
             header={
                 <Header
                     title="The last invoice is available"
-                    preamount="Some text (text)"
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle="Subtitle"
                 />
             }
             extra={<Placeholder />}
@@ -507,10 +499,6 @@ const headerSnippets: Array<Snippet> = [
             header={
                 <Header
                     title="The last invoice is available"
-                    preamount="Cuota mensual (IVA incluido)"
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle="Y esto es un subtitulo"
                 />
             }
             extra={<Placeholder />}
@@ -532,60 +520,9 @@ const headerSnippets: Array<Snippet> = [
             header={
                 <Header
                     title="The last invoice is available"
-                    preamount="Cuota mensual (IVA incluido)"
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle="Y esto es un subtitulo"
                 />
             }
             extra={<Placeholder />}
-        />
-        `,
-    },
-    {
-        group: 'Headers',
-        name: 'Header layout (rich text)',
-        code: `
-        <HeaderLayout
-            header={
-                <Header
-                    title="The last invoice is available"
-                    preamount={{
-                        text: "Example of line-through text",
-                        decoration: "line-through"
-                    }}
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle={{
-                        text: "Subtitle with secondary color",
-                        color: theme.colors.textSecondary
-                    }}
-                />
-            }
-        />
-        `,
-    },
-    {
-        group: 'Headers',
-        name: 'Header layout (rich text)(white)',
-        code: `
-        <HeaderLayout
-            isInverse={false}
-            header={
-                <Header
-                    title="The last invoice is available"
-                    preamount={{
-                        text: "Example of line-through text",
-                        decoration: "line-through"
-                    }}
-                    amount="60,44 €"
-                    button={<ButtonPrimary href="asdf">Action</ButtonPrimary>}
-                    subtitle={{
-                        text: "Subtitle with secondary color",
-                        color: theme.colors.textSecondary
-                    }}
-                />
-            }
         />
         `,
     },
@@ -672,7 +609,7 @@ const cardSnippets: Array<Snippet> = [
         name: 'MediaCard with Image',
         code: `
         <MediaCard
-            media={<Image src="https://picsum.photos/1200/1200" aspectRatio="16:9" />}
+            media={<Image src="https://picsum.photos/1200/1200" aspectRatio="16:9"/>}
             headline={<Tag type="promo">Headline</Tag>}
             pretitle="Pretitle"
             title="Title"
@@ -1502,6 +1439,108 @@ const skeletonSnippets = [
     },
 ];
 
+const heroSnippets = [
+    {
+        group: 'Hero',
+        name: 'White background',
+        code: `
+          <Hero
+            background="default"
+            media={
+              <Image
+                src="https://images.unsplash.com/photo-1604869515882-4d10fa4b0492?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1740&q=80"
+                aspectRatio="1:1"
+              />
+            }
+            headline={<Tag type="active">Novedad</Tag>}
+            pretitle="Brand Conecta Max"
+            title="Vuela con la Fibra 1Gb"
+            description="Para teletrabajar, ver series y películas y además, tener varios dispositivos conectados."
+            button={<ButtonPrimary fake>Lo quiero</ButtonPrimary>}
+            dataAttributes={{ testid: "hero" }}
+            desktopMediaPosition="right"
+          />
+      `,
+    },
+    {
+        group: 'Hero',
+        name: 'Color background',
+        code: `
+          <Hero
+            background="brand"
+            media={
+              <Image
+                src="https://images.unsplash.com/photo-1604869515882-4d10fa4b0492?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1740&q=80"
+                aspectRatio="16:9"
+              />
+            }
+            headline={<Tag type="active">Novedad</Tag>}
+            pretitle="Brand Conecta Max"
+            title="Vuela con la Fibra 1Gb"
+            description="Para teletrabajar, ver series y películas y además, tener varios dispositivos conectados."
+            button={<ButtonPrimary fake>Lo quiero</ButtonPrimary>}
+            dataAttributes={{ testid: "hero" }}
+            desktopMediaPosition="left"
+          />
+      `,
+    },
+    {
+        group: 'Hero',
+        name: 'Full height',
+        code: `
+          <Hero
+            background="default"
+            height="100vh"
+            media={
+              <Image
+                src="https://images.unsplash.com/photo-1604869515882-4d10fa4b0492?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1740&q=80"
+                aspectRatio="1:1"
+                height={isDesktopOrBigger ? undefined : "100%"}
+              />
+            }
+            headline={<Tag type="active">Novedad</Tag>}
+            pretitle="Brand Conecta Max"
+            title="Vuela con la Fibra 1Gb"
+            description="Para teletrabajar, ver series y películas y además, tener varios dispositivos conectados."
+            button={<ButtonPrimary fake>Lo quiero</ButtonPrimary>}
+            dataAttributes={{ testid: "hero" }}
+            desktopMediaPosition="left"
+          />
+      `,
+    },
+    {
+        group: 'Hero',
+        name: 'Slideshow',
+        code: `
+         <Slideshow
+            withBullets
+            inverseBullets={false}
+            items={Array.from({ length: 3 }, (_, idx) => (
+              <Hero
+                background="default"
+                media={
+                  <Image
+                    src="https://api.lorem.space/image/furniture?w=1500&h=1500&hash=8B7BCDC2"
+                    aspectRatio="16:9"
+                  />
+                }
+                headline={<Tag type="active">Headline</Tag>}
+                pretitle="Pretitle"
+                title={["Title", "Title2"][idx]}
+                description="This is a long description with a long text to see how this works"
+                extra={<Placeholder />}
+                button={<ButtonPrimary fake>Action</ButtonPrimary>}
+                buttonLink={<ButtonLink href="#">Link</ButtonLink>}
+                dataAttributes={{ testid: "hero" }}
+                desktopMediaPosition="right"
+              />
+            ))}
+          />
+
+      `,
+    },
+];
+
 export default [
     ...buttonSnippets,
     ...formSnippets,
@@ -1633,4 +1672,5 @@ export default [
     ...alertSnippets,
     ...tooltipSnippets,
     ...popoverSnippets,
+    ...heroSnippets,
 ].sort((s1, s2) => s1.group.localeCompare(s2.group)) as Array<Snippet>;