+++ src/__acceptance_tests__/__ssr_pages__/carousel.tsx
@@ -10,7 +10,12 @@ const CarouselTest = (): JSX.Element => (
                     headline={<Tag type="promo">Headline</Tag>}
                     title={'Card ' + idx}
                     description="Description"
-                    media={<Image src="https://api.lorem.space/image/watch?w=700&h=700" aspectRatio="16:9" />}
+                    media={
+                        <Image
+                            src="https://images.unsplash.com/photo-1575903013621-1387ce8caa74?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1740&q=80"
+                            aspectRatio="16:9"
+                        />
+                    }
                     button={
                         <ButtonPrimary small onPress={() => {}}>
                             Action