+++ src/header.tsx
@@ -26,25 +26,45 @@ type RichText = string | ({text: string} & OverridableTextProps);
 type HeaderProps = {
     pretitle?: RichText;
     title?: string;
-    preamount?: RichText;
-    amount?: string;
-    button?: RendersNullableElement<typeof ButtonPrimary>;
-    secondaryButton?: RendersNullableElement<typeof ButtonSecondary>;
-    subtitle?: RichText;
-    isErrorAmount?: boolean;
+    description?: string;
     dataAttributes?: DataAttributes;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    preamount?: RichText;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    amount?: string;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    button?: RendersNullableElement<typeof ButtonPrimary>;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    secondaryButton?: RendersNullableElement<typeof ButtonSecondary>;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    subtitle?: RichText;
+    /**
+     * @deprecated This field is deprecated, please use the extra slot in the HeaderLayout component instead.
+     */
+    isErrorAmount?: boolean;
 };
 
 export const Header: React.FC<HeaderProps> = ({
     pretitle,
     title,
+    description,
+    dataAttributes,
     preamount,
     amount,
     button,
     subtitle,
     isErrorAmount,
     secondaryButton,
-    dataAttributes,
 }) => {
     const {isTabletOrSmaller} = useScreenSize();
     const isInverse = useIsInverseVariant();
@@ -67,13 +87,18 @@ export const Header: React.FC<HeaderProps> = ({
 
     return (
         <Stack space={isTabletOrSmaller ? 24 : 32} dataAttributes={dataAttributes}>
-            {(title || pretitle) && (
+            {(title || pretitle || description) && (
                 <Box paddingRight={16}>
                     <Stack space={8}>
                         {pretitle && renderRichText(pretitle, {color: vars.colors.textPrimary})}
                         <Text6 role="heading" aria-level={2}>
                             {title}
                         </Text6>
+                        {description && (
+                            <Text3 regular color={vars.colors.textSecondary}>
+                                {description}
+                            </Text3>
+                        )}
                     </Stack>
                 </Box>
             )}
@@ -130,7 +155,11 @@ export const MainSectionHeader: React.FC<MainSectionHeaderProps> = ({title, desc
 type HeaderLayoutProps = {
     isInverse?: boolean;
     breadcrumbs?: RendersNullableElement<typeof NavigationBreadcrumbs>;
-    header: React.ReactNode; // intentionally not forced to RendersElement<typeof Header> to allow skeletons for example
+    /**
+     * Intentionally not forced to RendersElement<typeof Header> to allow skeletons for example
+     * The header is optional in order to allow webviews to delegate the header visualization to the surrounding native app.
+     */
+    header?: React.ReactNode;
     extra?: React.ReactNode;
     sideBySideExtraOnDesktop?: boolean;
     children?: void;
@@ -154,7 +183,7 @@ export const HeaderLayout: React.FC<HeaderLayoutProps> = ({
         >
             <OverscrollColor />
             {isTabletOrSmaller ? (
-                <Box paddingTop={32} paddingBottom={24}>
+                <Box paddingTop={header ? 32 : 0} paddingBottom={24}>
                     <Stack space={24}>
                         {header}
                         {extra}
@@ -175,19 +204,13 @@ export const HeaderLayout: React.FC<HeaderLayoutProps> = ({
                 </Box>
             ) : (
                 <Box paddingTop={breadcrumbs ? 16 : 48} paddingBottom={48}>
-                    <GridLayout
-                        template="6+6"
-                        left={
-                            <Stack space={24}>
-                                <Stack space={32}>
-                                    {breadcrumbs}
-                                    {header}
-                                </Stack>
-                                {extra}
-                            </Stack>
-                        }
-                        right={null}
-                    />
+                    <Stack space={isTabletOrSmaller ? 24 : 32}>
+                        <Stack space={32}>
+                            {breadcrumbs}
+                            {header}
+                        </Stack>
+                        {extra}
+                    </Stack>
                 </Box>
             )}
         </ResponsiveLayout>