+++ src/skins/blau.tsx
@@ -176,6 +176,12 @@ export const getBlauSkin: GetKnownSkin = () => {
             warningHigh: palette.blauYellow70,
             errorHigh: palette.blauRed70,
             promoHigh: palette.blauPurple,
+
+            successHighInverse: palette.blauGreen70,
+            warningHighInverse: palette.blauYellow70,
+            errorHighInverse: palette.blauRed70,
+            promoHighInverse: palette.blauPurple,
+            neutralMediumInverse: palette.grey5,
         },
         darkModeColors: {
             // BACKGROUNDS
@@ -261,16 +267,22 @@ export const getBlauSkin: GetKnownSkin = () => {
             textAppBarSelected: palette.grey2,
 
             // TAGS
-            successLow: applyAlpha(palette.white, 0.05),
-            warningLow: applyAlpha(palette.white, 0.05),
-            errorLow: applyAlpha(palette.white, 0.05),
-            promoLow: applyAlpha(palette.white, 0.05),
-            brandLow: applyAlpha(palette.white, 0.05),
+            successLow: palette.darkModeGrey6,
+            warningLow: palette.darkModeGrey6,
+            errorLow: palette.darkModeGrey6,
+            promoLow: palette.darkModeGrey6,
+            brandLow: palette.darkModeGrey6,
 
             successHigh: palette.blauGreen30,
             warningHigh: palette.blauYellow40,
             errorHigh: palette.blauRed40,
             promoHigh: palette.blauPurple30,
+
+            successHighInverse: palette.blauGreen70,
+            warningHighInverse: palette.blauYellow70,
+            errorHighInverse: palette.blauRed70,
+            promoHighInverse: palette.blauPurple,
+            neutralMediumInverse: palette.grey5,
         },
     };
 };