+++ src/skins/types.tsx
@@ -104,6 +104,7 @@ export type Colors = {
     neutralHigh: string;
     neutralLow: string;
     neutralMedium: string;
+    neutralMediumInverse: string;
     promo: string;
 
     // STATES
@@ -151,4 +152,8 @@ export type Colors = {
     warningHigh: string;
     errorHigh: string;
     promoHigh: string;
+    successHighInverse: string;
+    warningHighInverse: string;
+    errorHighInverse: string;
+    promoHighInverse: string;
 };